fastAPI/requirements.txt
setuptools
fastapi
uvicorn
python-dotenv
pymongo
pymongo[srv]